refactor: change tax rate type from double to decimal and update delivery cost calculation to include tax

+++ exercises/day19-North-Pole-Companies/csharp/NorthPole.Tests/InvoicePrinterTests.cs
@@ -75,7 +75,7 @@ public class InvoicePrinterTests
         {
             taxes[kvp.Key] = new TaxRate(
                 name: kvp.Value["name"].ToString(),
-                taxRate: kvp.Value["taxRate"].ToObject<double>(),
+                taxRate: kvp.Value["taxRate"].ToObject<decimal>(),
                 description: kvp.Value["description"].ToString());
         }
